drivers: gpio: Add more flags for GPIO's in devicetree
Currently only a subset of possible GPIO Flags is available inside the
devicetree. Add the remaining to ones from drivers/gpio.h into the gpio
devicetree header so that people can configure GPIO's inside the
devicetree.

+++ include/zephyr/drivers/gpio.h
@@ -246,7 +246,7 @@ typedef uint8_t gpio_pin_t;
  * The lower 8 bits are used for standard flags. The upper 8 bits are reserved
  * for SoC specific flags.
  */
-typedef uint16_t gpio_dt_flags_t;
+typedef uint32_t gpio_dt_flags_t;
 
 /**
  * @brief Provides a type to hold GPIO configuration flags.
